RequirementMachine: Split off PropertyUnification.cpp from PropertyMap.cpp
